delete gender test and create a new file for testing gender display
